LayoutTests/storage/indexeddb/setversion-blocked-by-versionchange-close-in-blocked-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test that open and setVersion are blocked if a VERSION_CHANGE transaction is running, and the connection is closed during the blocked event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-blocked-by-versionchange-close-in-blocked.html"
ver = 1
blockedEventFired = false
versionChangeComplete = false
errorEventFired = false
indexedDB.open(dbname)

h1OpenOnSuccess():
h1 = event.target.result
indexedDB.open(dbname)

h2OpenOnSuccess():
h2 = event.target.result
h1.setVersion(String(ver++))
h2.setVersion(String(ver++))

h2OnVersionChange():
old = 1
new = "1"

h1SetVersionOnBlocked():
blockedEventFired = true

h1OnVersionChange():
old = 1
new = "2"

h2SetVersionOnBlocked():
h2.close()
indexedDB.open(dbname)

h2SetVersionOnError():
errorEventFired = true

h1SetVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

h3OpenOnSuccess():
PASS blockedEventFired is true
PASS versionChangeComplete is true
PASS errorEventFired is true
PASS successfullyParsed is true

TEST COMPLETE
